% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_preProcessing.R
\name{abstractReduction}
\alias{abstractReduction}
\title{Reduce the size of your PubMed abstract texts}
\usage{
abstractReduction(data, fraction)
}
\arguments{
\item{data}{PubMed entry data (see 'pubRetrieve')}

\item{fraction}{The decimal represenitng the fraction of each abstract to trim. We recommend (0.2) or 20%}
}
\description{
Trim abstracts from the head and remove any entries without abstract text data.
}
\examples{
abstractReduction(data = abstract_data, fraction = 0.2)
}
\keyword{Text}
\keyword{reduction}
