% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCAF.R
\name{add_underscores_to_compound_words}
\alias{add_underscores_to_compound_words}
\title{Replace bioentity noun phrase spacing with underscores to preserve structure}
\usage{
add_underscores_to_compound_words(character_list)
}
\arguments{
\item{character_list}{List of named entities in character type}
}
\description{
This function is not intended for users
}
