% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1_AOP.R
\name{AOP_XML_children_organizer}
\alias{AOP_XML_children_organizer}
\title{Converts AOP-wiki XML files into R data frames for further processing}
\usage{
AOP_XML_children_organizer(doc)
}
\arguments{
\item{doc}{XML file downloaded from AOP-wiki}
}
\description{
This function takes XML and converts
}
\examples{
Word_Cleaner(data = association_rules_df)
}
\keyword{AOP}
\keyword{AOP-wiki,}
\keyword{Classification,}
