% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCAF.R
\name{bioCAF}
\alias{bioCAF}
\title{Mine bioconcept associations by combining PubTator and KAF}
\usage{
bioCAF(data, choice, support, venv_path, lang_model)
}
\arguments{
\item{data}{Generated using PubTator}

\item{choice}{Control which associations to mine and retrieve for analysis: "All", Chemical", "Disease", "Gene", "Systox" (Chemical and disease associations)}

\item{support}{To control computational load, support is modifiable according to each query and user needs. If you are generating too many results for your liking, you can increase this support threshold which represents the percentage of (=<2000) abstracts}

\item{venv_path}{REQUIRED: The path to your python venv. (see 'https://github.com/dandycodingpipe/KAFtool' for additional information)}

\item{lang_model}{REQUIRED: The spaCy language model installed in your venv}
}
\description{
Bioconcepts associations this function can produce are chemical-gene, gene-gene, or disease-chemical/disease-gene.
}
\examples{
bioCAF(data, "Disease", 0.001, "C:/Users/Chris/OneDrive/2023/Systox/venvJune19", "en_core_web_lg")
}
\keyword{KAF,}
\keyword{PubMed}
\keyword{PubTator,}
