% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_infoRetrieval.R
\name{europepmc_retrieval}
\alias{europepmc_retrieval}
\title{The Europe PMC search and retrieve function used for sourcing abstracts.}
\usage{
europepmc_retrieval(query, retmax)
}
\arguments{
\item{query}{Your typical PubMed query. Optimizing your query using the proper pubmed syntax improves results!}

\item{retmax}{Define how many entries you'd like to access for a query. Keep it between 500-1000 for fastest results. (Limit is around 20,000)}
}
\description{
This function allows you to query and access key entry data from European PubMed central.
}
\examples{
europepmc_retrieval(Query = "Vape smoking AND toxicity", retmax = 750)
}
\keyword{EuropePMC,}
\keyword{PMC,}
\keyword{Pubmed,}
\keyword{Query,}
\keyword{Retrieval}
