% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_visualization.R
\name{KAFviewer}
\alias{KAFviewer}
\title{The KAF visualization suite}
\usage{
KAFviewer(rules, viz, class)
}
\arguments{
\item{rules}{The data frame of rules you want to classify}

\item{viz}{The visualization option ("df" or "bar")}

\item{class}{The classification display option ("all","bme","anatomy", "organism","diseases","chemicals","techniques","other","unlab")}
}
\description{
This function allows you to display rules as a bargraph or as a dataframe according to the classification of choice.
}
\examples{
ruleViewer(viz <- ruleViewer(classified_rules, "bar", "bme"))
}
\keyword{ARM,}
\keyword{Association-rule}
\keyword{MeSH,}
\keyword{Post}
\keyword{classification,}
\keyword{mining,}
\keyword{processing,}
\keyword{visualization}
