% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1_AOP.R
\name{matchAOP}
\alias{matchAOP}
\alias{KAFxAOP}
\title{Compare KAF results to the AOP-wiki database to find notable key events, stressors, and biological processes present in your PubMed query.}
\usage{
KAFxAOP(sample, AOPdownload)
}
\arguments{
\item{rules}{Association rule dataframe that was mined using Abstract_ARM or easyKAF. Limit this parameter to 5000-20000 rules as deduplication is very computationally demanding.}
}
\description{
(In development) Uses locality-sensitive hashing to efficiently cluster and calculate the Jaccard similarity between mined rules and AOP-wiki values/classes. Values below 60% similarity are not considered matches.
}
\examples{
KAFxAOP(rules)
}
\keyword{AOP}
\keyword{Hashing,}
\keyword{Jaccard}
\keyword{LSH,}
\keyword{Locality}
\keyword{Similarity,}
\keyword{deduplication,}
\keyword{fuzzy}
\keyword{hashing,}
\keyword{matching,}
\keyword{sensitive}
