% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2_classificationMeSH.R
\name{matchMeSH}
\alias{matchMeSH}
\title{The KAF classification work-horse}
\usage{
matchMeSH(raw_rules, removal)
}
\arguments{
\item{raw_rules}{The data frame of rules you want to classify}

\item{removal}{A list of words that crash the classifier. These words are removed and not classified.}
}
\description{
This function handles the classification suite of functions and outputs cleaned and classified association rules
}
\examples{
matchMeSH(rules, removal = c("ml(-1","sub>2</sub","study","lead","±", "°", "confidence", "-", "\%", "β", ">", "sub>50</sub","müllerian", "#"))
}
\keyword{ARM,}
\keyword{Association-rule}
\keyword{MeSH}
\keyword{Post}
\keyword{classification,}
\keyword{mining,}
\keyword{processing,}
