% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2_classificationMeSH.R
\name{MeSH_filter}
\alias{MeSH_filter}
\title{This function simply rebuilds the association rules that the MeSH cleaner had removed.}
\usage{
MeSH_filter(MeSH_data)
}
\arguments{
\item{MeSH_data}{The data frame of rules you want to re-structure}
}
\description{
Used for MeSH_Cleaner.
}
\examples{
MeSH_filter(MeSH_data)
}
\keyword{ARM,}
\keyword{Association-rule}
\keyword{MeSH}
\keyword{Post}
\keyword{classification,}
\keyword{mining,}
\keyword{processing,}
