% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2_classificationMeSH.R
\name{MeSH_Mapper}
\alias{MeSH_Mapper}
\title{A function that communicates with the Medical Subject Headings (MeSH) RDF API to classify consequenets (RHS) from mined association rules.}
\usage{
MeSH_Mapper(word, removal)
}
\arguments{
\item{word}{The data frame of rules you want to classify}

\item{removal}{A list of words you'd like to remove from the classification task (typically because they crash the process)}
}
\description{
This function is the classification workhouse which outputs the same association-rule dataframe you input, but with an additional column dedicated to the classification values.
}
\examples{
MeSH_Mapper(rules, removal = c("ml(-1","sub>2</sub","study","lead","±", "°", "confidence", "-", "\%", "β", ">", "sub>50</sub","müllerian", "#"))
}
\keyword{ARM,}
\keyword{Association-rule}
\keyword{Post}
\keyword{classification}
\keyword{mining,}
\keyword{processing,}
