% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_infoRetrieval.R
\name{pubmed_retrieval}
\alias{pubmed_retrieval}
\title{PubMed multi-entry and retrieval function}
\usage{
pubmed_retrieval(query, retmax)
}
\arguments{
\item{query}{Your typical PubMed query. Optimizing your query using the proper pubmed syntax improves results!}

\item{retmax}{Define how many entries you'd like to access for a query. Keep it between 500-1000 for fastest results. (Limit is around 2000)}
}
\description{
This function allows you to query and access key entry data from PubMed.
}
\examples{
pubmed_retrieval(Query = "Vape smoking AND toxicity", retmax = 750)
}
\keyword{PMC,}
\keyword{Pubmed,}
\keyword{Query,}
\keyword{Retrieval}
