% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_machineLearning.R
\name{pubMine}
\alias{pubMine}
\title{Mine association-rules on parsed PubMed abstract data}
\usage{
pubMine(data, min_supp, min_conf, min_p)
}
\arguments{
\item{data}{Parsed data (See pubParse)}

\item{min_supp}{Define the minimum support threshold for the rules (e.g 0.01 or 0.10)}

\item{min_conf}{Define the minimum confidence threshold for the rules (e.g 0.50 or 0.90)}

\item{min_p}{Define the minimum p-value threshold for the rules (e.g 0.05 or 0.0.005)}
}
\description{
The output is an R dataframe containing association rules and associated metrics. Outputs can vary greatly depending on the parsing method.
}
\examples{
pubMine(data <- data_from_Text_Parser, min_supp = 0.01, min_conf = 0.75, min_p = 0.005)
}
\keyword{Processing}
\keyword{algorithm,}
\keyword{arules,}
\keyword{association-rules}
\keyword{mining,}
\keyword{statistics,}
