% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_infoRetrieval.R
\name{pubRetrieve}
\alias{pubRetrieve}
\title{Extract PubMed or Europe PMC entry information for any query}
\usage{
pubRetrieve(query, size, database)
}
\arguments{
\item{query}{Your typical PubMed query. Optimizing your query using the proper PubMed or Europe PMC syntax improves results.}

\item{size}{An integer representing the number of entries to retrieve. PubMed is limited to about 2000 articles, while Europe PMC can pull around 20,000.}

\item{database}{Define the PubMed database to retrieve from: "pubmed" or "pmc".}
}
\description{
Combines functionality from two R-packages: "easyPubMed" and "europepmc" for reliable PubMed entry data extraction.
}
\examples{
pubRetrieve(query = "Vape smoking AND toxicity", size = 2000, database = "pubmed")
}
\keyword{Europe}
\keyword{Information}
\keyword{PMC}
\keyword{PubMed,}
\keyword{retrieval,}
