% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_pubTator.R
\name{PubTator}
\alias{PubTator}
\title{Retrieve bio-entities in PubMed data using PubTator}
\usage{
PubTator(query)
}
\arguments{
\item{query}{Input a PubMed query (see https://pubmed.ncbi.nlm.nih.gov/help/#how-do-i-search-pubmed)}
}
\description{
This function uses PubTator-annotated articles to retrieve pre-annotated articles. These results should be passed onto bioKAF for further processing (See bioKAF).
}
\examples{
PubTator("dioxin toxicity","C:/Users/Chris/OneDrive/2023/Systox/venvJune19", "en_core_web_lg" )
}
\keyword{KAF,}
\keyword{PubMed}
\keyword{PubTator,}
