% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_preProcessing.R
\name{rm_negation}
\alias{rm_negation}
\title{The function removes sentences possessing negations. Used in Text_Parser()}
\usage{
rm_negation(Abstract_Parse)
}
\arguments{
\item{Abstract_Parse}{Input your parsed abstracts. (typically obtained from the spacy_parse function)}
}
\description{
This function removes negations in already-parsed abstract corpus.
}
\examples{
rm_negation(Abstract_Parse = parsed_abstracts_from_spacy_parse)
}
\keyword{NLP,}
\keyword{Text}
\keyword{language}
\keyword{modifier}
\keyword{natural}
\keyword{negation}
\keyword{processing,}
\keyword{reduction,}
