% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2_classificationMeSH.R
\name{Rule_Concatenator}
\alias{Rule_Concatenator}
\title{Clean association rules}
\usage{
Rule_Concatenator(rules)
}
\arguments{
\item{rules}{The data frame of rules you want to concatenate}
}
\description{
Before fuzzy matching we have to pre-process the rules by concatenating them into a single string
}
\examples{
Rule_Concatenator(rules)
}
\keyword{ARM,}
\keyword{Association-rule}
\keyword{MeSH}
\keyword{Post}
\keyword{classification,}
\keyword{mining,}
\keyword{processing,}
