% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Exposome_Network.R
\name{systox}
\alias{systox}
\title{Bio-concept co-occurence network visualization}
\usage{
systox(data, style)
}
\arguments{
\item{data}{stricly-bioCAF association rules data-frame}

\item{style}{The network layout style. We recommend "layout_nicely" or "layout_in_circle" (For additional styles see https://igraph.org/r/doc/layout_.html)}

\item{class}{The classification display option ("all","bme","anatomy", "organism","diseases","chemicals","techniques","other","unlab")}
}
\description{
This function allows you visualize the literature landscape as an interactive network and explore how bioconcepts directly or indirectly relate to one another.
}
\examples{
concepts <- PubTator("tobacco smoke toxicity")
results <- bioCAF(concepts, "all", 0.003, "C:/Users/Chris/OneDrive/2023/Systox/venvJune19", "en_core_web_lg")
viz <- systox(results, "layout_nicely")

#export option as html
library(htmlwidgets)
saveWidget(viz, file = "bioTM_hyperthyroidism.html")
}
\keyword{Network,}
\keyword{co-occurence}
\keyword{exposome,}
\keyword{exposomics}
\keyword{network,}
