% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2_classificationMeSH.R
\name{Word_Cleaner}
\alias{Word_Cleaner}
\title{A crude consequent or RHS remover function for ambigious rules.}
\arguments{
\item{data}{Input your mined associations rules.}
}
\description{
This function removes the bracketing present in association-rules (e.g {consequent}) for better post-processing
}
\examples{
Word_Cleaner(data = association_rules_df)
}
\keyword{ARM,}
\keyword{Association-rule}
\keyword{Post}
\keyword{classification}
\keyword{mining,}
\keyword{processing,}
